epoll

+++ longServer/src/ThreadPool.h
@@ -21,7 +21,7 @@ private:
 
     SyncQueue<Task> tTasks;								//任务队列
     std::vector<std::shared_ptr<std::thread>> tThreads; //线程对象
-    unsigned short nThreadNum;							//线程数
+    unsigned int nThreadNum;							//线程数
 
     std::mutex tMutex;					//锁
     std::condition_variable tCondition; //条件变量
